fix parsing erors
